rename commands package to command
