+++ internal/command/config.go
@@ -1,4 +1,4 @@
-package commands
+package command
 
 import (
 	"encoding/json"
